LiveSPANNEDML/requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.0
numpy>=1.20.0
torch>=1.10.0
transformers>=4.18.0
pytube>=12.0.0
openai-whisper>=1.0.0
matplotlib>=3.5.0
